requirements.txt
psycopg2-binary
qdrant_client
firebase-admin
torch
numpy
opencv-python